+++ extensions/elation/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Elation Changelog
 
-## 2023-08-21
+## September 4, 2023
+
+- rename `onCreatePatient` webhook to `patientCreatedOrUpdated`
+- filter non `saved` actions for this webhook
+
+## August 21, 2023
 
 - make non-visit note text a large input (`StringType.TEXT`)